Moved JWT generation into separate module
